agregando Security

+++ routes/photo.py
@@ -1,11 +1,13 @@
+
 from fastapi import APIRouter, Depends, Security
 from schemas.photo import PhotoBase, PhotoDisplay, PhotoDelete
 from typing import List
 from config.database import get_db
 from config import db_photo
 
-
 router = APIRouter(prefix='/api')
+
+
 
 @router.get('/', response_model=List[PhotoDisplay])
 async def get_photos(db = Depends(get_db)):
@@ -22,4 +24,6 @@ async def update_photo(id: str, request: PhotoBase, db = Depends(get_db)):
 
 @router.delete('/{id}')
 async def delete_photo(id: str, request: PhotoDelete, db = Depends(get_db)):
+    if (request.userCode != USER_CODE):
+        return {'error': 'Your User code is not correct'}
     return db_photo.delete_one(id, request, db)
